+++ packages/categorize/src/categorize/choice.jsx
@@ -2,7 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import classNames from 'classnames';
-import { DragSource } from 'react-dnd';
+import { DragSource, DndProvider } from 'react-dnd';
+import HTML5Backend from 'react-dnd-html5-backend';
 import { uid } from '@pie-lib/pie-toolbox/drag';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
@@ -95,13 +96,33 @@ export class Choice extends React.Component {
     connectDragSource: PropTypes.func.isRequired,
   };
 
+  componentDidMount() {
+    if (this.ref) {
+      this.ref.addEventListener('touchstart', this.handleTouchStart, { passive: false });
+    }
+  }
+
+  componentWillUnmount() {
+    if (this.ref) {
+      this.ref.removeEventListener('touchstart', this.handleTouchStart);
+    }
+  }
+
+  handleTouchStart = (e) => {
+    e.preventDefault();
+  };
+
   render() {
     const { connectDragSource, id, content, disabled, isDragging, correct, extraStyle } = this.props;
 
     return connectDragSource(
-      <div style={{ margin: '4px', ...extraStyle }}>
-        <Styled id={id} content={content} disabled={disabled} correct={correct} isDragging={isDragging} />
-      </div>,
+        <div
+            style={{ margin: '4px', ...extraStyle }}
+            ref={(ref) => (this.ref = ref)}
+            draggable={!disabled}
+        >
+          <Styled id={id} content={content} disabled={disabled} correct={correct} isDragging={isDragging} />
+        </div>,
     );
   }
 }
@@ -120,14 +141,14 @@ export const spec = {
     return out;
   },
   endDrag: (props, monitor) => {
-    if (!monitor.didDrop()) {
-      const item = monitor.getItem();
-      if (item.categoryId) {
-        log('wasnt droppped - what to do?');
-        props.onRemoveChoice(item);
+      if (!monitor.didDrop()) {
+        const item = monitor.getItem();
+        if (item.categoryId) {
+          log('wasnt droppped - what to do?');
+          props.onRemoveChoice(item);
+        }
       }
     }
-  },
 };
 
 const DraggableChoice = DragSource(
@@ -139,4 +160,10 @@ const DraggableChoice = DragSource(
   }),
 )(Choice);
 
-export default uid.withUid(DraggableChoice);
+const DraggableChoiceWithProvider = (props) => (
+    <DndProvider backend={HTML5Backend}>
+      <DraggableChoice {...props} />
+    </DndProvider>
+);
+
+export default uid.withUid(DraggableChoiceWithProvider);