+++ packages/extended-text-entry/src/main.jsx
@@ -180,6 +180,9 @@ export class Main extends React.Component {
               table: {
                 disabled: true,
               },
+              textAlign: {
+                disabled: true,
+              }
             }}
             languageCharactersProps={languageCharactersProps}
           />