+++ packages/boilerplate-item-type/controller/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.17](https://github.com/pie-framework/pie-elements/compare/@pie-element/boilerplate-item-type-controller@3.1.16...@pie-element/boilerplate-item-type-controller@3.1.17) (2025-03-03)
+
+
+### Bug Fixes
+
+* update pie-lib pie-toolbox PD-4795 ([c0bc482](https://github.com/pie-framework/pie-elements/commit/c0bc48285a67aaaf6f4b81057ced62bc57ee978d))
+
+
+
+
+
+## [3.1.16](https://github.com/pie-framework/pie-elements/compare/@pie-element/boilerplate-item-type-controller@3.1.15...@pie-element/boilerplate-item-type-controller@3.1.16) (2025-02-27)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-4820, PD-4766 ([7ccdb40](https://github.com/pie-framework/pie-elements/commit/7ccdb405614f1c310efe530d97c46fe324c0a669))
+
+
+
+
+
 ## [3.1.15](https://github.com/pie-framework/pie-elements/compare/@pie-element/boilerplate-item-type-controller@3.1.14...@pie-element/boilerplate-item-type-controller@3.1.15) (2025-02-19)