+++ packages/categorize/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [9.11.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@9.10.8...@pie-element/categorize@9.11.0) (2025-03-12)
+
+
+### Bug Fixes
+
+* **categorize:** remove test as logic was removed regarding small movement on token PD-4866 ([d8619fa](https://github.com/pie-framework/pie-elements/commit/d8619fa9f23a52e3dd6a4cbef1fd123197c1015b))
+* **multiple-choice, categorize, ditb, hotspot, ica:** remove audio events listeners on disconnectedCallback PD-4852 ([35937aa](https://github.com/pie-framework/pie-elements/commit/35937aa5a790e24ead17b95561c368056b5a14f8))
+
+
+### Features
+
+* **categorize:** add customAudioButton PD-4795 ([d712cff](https://github.com/pie-framework/pie-elements/commit/d712cff7b3c0c26219768a12f0419a9d3629daf3))
+
+
+
+
+
+## [9.10.8](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@9.10.7...@pie-element/categorize@9.10.8) (2025-03-03)
+
+
+### Bug Fixes
+
+* prevent autoplay logic when autoplayAudioEnabled is missing or false PD-4847 ([211ae94](https://github.com/pie-framework/pie-elements/commit/211ae940a238950ae603b54cb6c294d7e68cd2cb))
+* update pie-lib pie-toolbox PD-4795 ([c0bc482](https://github.com/pie-framework/pie-elements/commit/c0bc48285a67aaaf6f4b81057ced62bc57ee978d))
+
+
+
+
+
+## [9.10.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@9.10.6...@pie-element/categorize@9.10.7) (2025-02-27)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-4820, PD-4766 ([7ccdb40](https://github.com/pie-framework/pie-elements/commit/7ccdb405614f1c310efe530d97c46fe324c0a669))
+* **categorize:** fix tests PD-4763 ([2a88abb](https://github.com/pie-framework/pie-elements/commit/2a88abb1f62072af281d60060818d13aa3d1ed64))
+* **categorize:** on touch devices on long press do not remove choice PD-4763 ([126c19a](https://github.com/pie-framework/pie-elements/commit/126c19a6e384e29367532e842e5bdcb23890413c))
+
+
+
+
+
 ## [9.10.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@9.10.5...@pie-element/categorize@9.10.6) (2025-02-25)
 
 **Note:** Version bump only for package @pie-element/categorize