+++ packages/categorize/controller/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.7.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-controller@7.6.4...@pie-element/categorize-controller@7.7.0) (2025-03-12)
+
+
+### Features
+
+* **categorize:** add customAudioButton PD-4795 ([d712cff](https://github.com/pie-framework/pie-elements/commit/d712cff7b3c0c26219768a12f0419a9d3629daf3))
+
+
+
+
+
+## [7.6.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-controller@7.6.3...@pie-element/categorize-controller@7.6.4) (2025-03-03)
+
+
+### Bug Fixes
+
+* update pie-lib pie-toolbox PD-4795 ([c0bc482](https://github.com/pie-framework/pie-elements/commit/c0bc48285a67aaaf6f4b81057ced62bc57ee978d))
+
+
+
+
+
+## [7.6.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-controller@7.6.2...@pie-element/categorize-controller@7.6.3) (2025-02-27)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-4820, PD-4766 ([7ccdb40](https://github.com/pie-framework/pie-elements/commit/7ccdb405614f1c310efe530d97c46fe324c0a669))
+
+
+
+
+
 ## [7.6.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-controller@7.6.1...@pie-element/categorize-controller@7.6.2) (2025-02-19)