+++ packages/hotspot/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.7.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@7.6.7...@pie-element/hotspot@7.7.0) (2025-03-12)
+
+
+### Bug Fixes
+
+* **multiple-choice, categorize, ditb, hotspot, ica:** remove audio events listeners on disconnectedCallback PD-4852 ([35937aa](https://github.com/pie-framework/pie-elements/commit/35937aa5a790e24ead17b95561c368056b5a14f8))
+
+
+### Features
+
+* **hotspot:** add customAudioButton PD-4795 ([16177af](https://github.com/pie-framework/pie-elements/commit/16177afaa11beffcea26798b64a67eebb5c6833b))
+
+
+
+
+
+## [7.6.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@7.6.6...@pie-element/hotspot@7.6.7) (2025-03-03)
+
+
+### Bug Fixes
+
+* prevent autoplay logic when autoplayAudioEnabled is missing or false PD-4847 ([211ae94](https://github.com/pie-framework/pie-elements/commit/211ae940a238950ae603b54cb6c294d7e68cd2cb))
+* update pie-lib pie-toolbox PD-4795 ([c0bc482](https://github.com/pie-framework/pie-elements/commit/c0bc48285a67aaaf6f4b81057ced62bc57ee978d))
+
+
+
+
+
+## [7.6.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@7.6.5...@pie-element/hotspot@7.6.6) (2025-02-27)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-4820, PD-4766 ([7ccdb40](https://github.com/pie-framework/pie-elements/commit/7ccdb405614f1c310efe530d97c46fe324c0a669))
+
+
+
+
+
 ## [7.6.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@7.6.4...@pie-element/hotspot@7.6.5) (2025-02-25)
 
 **Note:** Version bump only for package @pie-element/hotspot