+++ packages/hotspot/controller/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot-controller@5.3.4...@pie-element/hotspot-controller@5.4.0) (2025-03-12)
+
+
+### Features
+
+* **hotspot:** add customAudioButton PD-4795 ([16177af](https://github.com/pie-framework/pie-elements/commit/16177afaa11beffcea26798b64a67eebb5c6833b))
+
+
+
+
+
+## [5.3.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot-controller@5.3.3...@pie-element/hotspot-controller@5.3.4) (2025-03-03)
+
+
+### Bug Fixes
+
+* update pie-lib pie-toolbox PD-4795 ([c0bc482](https://github.com/pie-framework/pie-elements/commit/c0bc48285a67aaaf6f4b81057ced62bc57ee978d))
+
+
+
+
+
+## [5.3.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot-controller@5.3.2...@pie-element/hotspot-controller@5.3.3) (2025-02-27)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-4820, PD-4766 ([7ccdb40](https://github.com/pie-framework/pie-elements/commit/7ccdb405614f1c310efe530d97c46fe324c0a669))
+
+
+
+
+
 ## [5.3.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot-controller@5.3.1...@pie-element/hotspot-controller@5.3.2) (2025-02-19)